New Picture Comparer
Changed to newer version of picture comparer

+++ src/UiComponents/UiMainPanel.java
@@ -42,7 +42,7 @@ public class UiMainPanel extends JPanel {
         write(String.format("%s%n", msg));
     }
 
-    public void updateTray(int total, int processed, int duplicates) {
+    public void updateTray(long total, long processed, long duplicates) {
         uiTray_.update(total, processed, duplicates);
     }
